+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Change Log
 * Minor changes to view_portal_object utility script.
 * Minor changes to validators hooks in structured_data.
 * Added portal_utils.Portal.get_version method.
+* Minor fix in misc_utils.format_duration.
 
 
 8.14.3